+++ src/test/resources/pgdiff_test_files/add_rule_new.sql
@@ -1,0 +1,11 @@
+create schema x;
+
+create table x.table1
+(
+    x int
+);
+
+CREATE RULE hello_world_added AS
+    ON UPDATE TO x.table1 DO
+    NOTIFY hello_added;
+