+++ src/test/resources/pgdiff_test_files/add_rule_original.sql
@@ -1,0 +1,10 @@
+create schema x;
+
+create table x.table1
+(
+    x int
+);
+
+CREATE RULE notify_me AS
+    ON UPDATE TO x.table1 DO
+    NOTIFY hello;