+++ src/test/resources/pgdiff_test_files/drop_function_args2_diff.sql
@@ -1,2 +1,1 @@
-
-DROP FUNCTION multiply_numbers(integer, integer);
+DROP FUNCTION IF EXISTS "public".multiply_numbers(integer, integer);