+++ src/test/resources/pgdiff_test_files/drop_unlogged_table_diff.sql
@@ -1,4 +1,1 @@
-
-SET search_path = radstg, pg_catalog;
-
-DROP TABLE IF EXISTS asset_country_weight;
+DROP TABLE IF EXISTS radstg.asset_country_weight;