+++ src/test/resources/pgdiff_test_files/drop_view_diff.sql
@@ -1,2 +1,1 @@
-
-DROP VIEW IF EXISTS testview;
+DROP VIEW IF EXISTS "public".testview;